saluki_health

Struct HealthRegistry

pub struct HealthRegistry { /* private fields */ }

A registry of components and their health.

HealthRegistry is responsible for tracking the health of all registered components, by storing both their readiness, which indicates whether or not they are initialized and generally ready to process data, as well as probing their liveness, which indicates if theyโre currently responding, or able to respond, to requests.

Telemetry

The health registry emits some internal telemetry about the status of registered components. In particular, three metrics are emitted:

health.component_ready: whether or not a component is ready (gauge, 0 for not ready, 1 for ready)

health.component_alive: whether or not a component is alive (gauge, 0 for not alive/unknown, 1 for alive, -1 for dead)

health.component_liveness_latency_secs: the response latency of the component for liveness probes (histogram, in seconds)

All metrics have a component_id tag that corresponds to the name of the component that was given when registering it.

Implementations

impl HealthRegistry

pub fn new() -> Self

Creates an empty registry.

pub fn register_component<S: Into<MetaString>>(&self, name: S) -> Option<Health>

Registers a component with the registry.

A handle is returned that must be used by the component to set its readiness as well as respond to liveness probes. See Health::mark_ready, Health::mark_not_ready, and Health::live for more information.

pub fn api_handler(&self) -> HealthAPIHandler

Gets an API handler for reporting the health of all components.

This handler exposes routes for querying the readiness and liveness of all registered components. See HealthAPIHandler for more information about routes and responses.

pub fn all_ready(&self) -> bool

Returns true if all components are ready.

pub async fn spawn(self) -> Result<JoinHandle<()>, GenericError>

Spawns the health registry runner, which manages the scheduling and collection of liveness probes.

Errors

If the health registry has already been spawned, an error will be returned.